Allow getting input from stdin
If the input list is empty and the machine requests input, we now block
waiting for further input from stdin. (We used to just crash....)
Accordingly the function run is moved into the IO Monad.

I also renamed State Input to GetIn to remove name clash with Input Data
type.