Change to a multi-threaded approach for day 7 pt 1
For part 1 I don't think there's much advantage to this compared to
simply sequentially looping over the machines and taking each one's
output as input for the next. Each machine will block early in execution
waiting for its single input, anyway. But a threaded approach is natural
for part 2, so this should help prepare for that.