Make machine print outputs as it runs
...instead of waiting until the end of the simulation and then flushing
an output buffer. This should make it faster for several interacting
machines to grab and use each others' output as they all run (e.g., day
7 part 2).

We also move the input list outside of the computer and include it as an
argument to the run function instead. It was probably a mistake from the
start to try to "fake" input and output as states of the computer.