Clean up type synonyms and function type signatures
Basically removes hardcoding of Ints. This should make the intended use
of those Ints clearer and also make it much easier to change the word
size throughout if that is necessary (it should be sufficient to change
the definition of IWord and change the Program type from an IntMap to a
normal Map).